fix: `cat:swissCatNumber`
Don't make that property mandatory, since it is okay if a Chemical has a `cat:casNumber` instead. We can make the demand stronger later if needed. As we have been told, the `cat:casNumber` is not always there in that cases a `cat:swissCasNumber` is provided. In principle it can always be provided, but the current data don't always have it.